% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{base_map}
\alias{base_map}
\alias{add_people}
\alias{add_tracts}
\title{Plot census tract boundaries and demographic data on a leaflet map}
\usage{
base_map()

add_people(lmap, year_id, people_data)

add_tracts(lmap, year_id, tract_data)
}
\arguments{
\item{lmap}{A \code{\link[leaflet]{leaflet}} map object}

\item{year_id}{A four-digit year}

\item{people_data}{A dataframe containing locations to place dots representing people, output of create_dots function}

\item{tract_data}{A dataframe containing shapefiles to plot census tract boundaries for each year, output of get_data_wide function}
}
\description{
Plot census tract boundaries and demographic data on a leaflet map
}
\examples{
\dontrun{
add_people(base_map(), 1915)
add_people(base_map(), 1960)
add_people(base_map(), 1975)
# should see same color mapping
add_people(base_map(), 2011)
# shouldn't see both sets of points
add_people(add_people(base_map(), 1975), 2011)
}
\dontrun{
add_tracts(base_map(), 1960)
add_tracts(base_map(), 1975)
add_tracts(add_tracts(base_map(), 1960), 2010)
}
}
