% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_data_long}
\alias{boston_data_long}
\title{US Census data for Boston, MA (Suffolk County) in long format}
\format{
A tibble with 16319 rows and 11 columns
\describe{
\item{GISJOIN}{The census tract}
\item{STATE}{The state the county is located in}
\item{COUNTY}{The County}
\item{variable}{Original (IPUMS) variable name that categorizes race}
\item{n}{The number of people of in the race category in that tract that year}
\item{num_people}{The total number of people in that tract for that year}
\item{pct_people}{The percent that the race category is in that tract that year}
\item{year}{Census year}
\item{x}{A latitude longitude point}
\item{census_label}{Census label corresponding to variable name}
\item{race_label}{Race label manually assigned to variable for the purpose of displaying on the map}
\item{is_hispanic}{Logical telling whether or not individuals in the labeled group are Hispanic}
}
}
\source{
IPUMS: data can be accessed at \url{https://www.ipums.org/}
}
\usage{
boston_data_long
}
\description{
A dataset for the population of Boston, MA in long format
}
\keyword{datasets}
