% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorPeople}
\alias{colorPeople}
\title{Color palette to be shared across maps}
\usage{
colorPeople(palette = "Zissou 1")
}
\arguments{
\item{palette}{Name of color palette passed to
\code{\link[grDevices]{hcl.colors}}}
}
\description{
Color palette to be shared across maps
}
