% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_people.R
\name{create_dots}
\alias{create_dots}
\title{Generate dataframe with locations of dots representing people on map}
\usage{
create_dots(data, num_people = 100)
}
\description{
Generate dataframe with locations of dots representing people on map
}
