% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_data_long}
\alias{filter_data_long}
\title{Generate data frame with demographic information for 1950-2020 for specified region}
\usage{
filter_data_long(data, state, county)
}
\arguments{
\item{state}{A character string specifying the name of the desired state}

\item{county}{A character string specifying the name of the desired county}

\item{path}{A character string specifying the path where data is located}
}
\description{
Generate data frame with demographic information for 1950-2020 for specified region
}
