% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{get_data_long}
\alias{get_data_long}
\alias{filter_data_wide}
\title{Read in and load tracts_long_all}
\usage{
get_data_long()

filter_data_wide(data, state, county)
}
\description{
Read in and load tracts_long_all

Generate data frame used to map demographics
}
