% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{madison_data_wide}
\alias{madison_data_wide}
\title{US Census data for Madison, NY (Madison County) in wide format}
\format{
A tibble with 7 rows and 2 columns
\describe{
\item{year}{Census year}
\item{tract_data}{An sf list with the data needed to plot the data on the map}
}
}
\source{
IPUMS: data can be accessed at \url{https://www.ipums.org/}
}
\usage{
madison_data_wide
}
\description{
A dataset for the population of Madison, NY
}
\keyword{datasets}
