% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sanfrancisco_sample}
\alias{sanfrancisco_sample}
\title{US Census data for San Francisco, CA (San Francisco County)}
\format{
A tibble with 60125 rows and 6 columns
\describe{
\item{year}{Census year}
\item{variable}{Original (IPUMS) variable name that categorizes race}
\item{x}{A latitude longitude point}
\item{census_label}{Census label corresponding to variable name}
\item{race_label}{Race label manually assigned to variable for the purpose of displaying on the map}
\item{is_hispanic}{Logical telling whether or not individuals in the labeled group are Hispanic}
}
}
\source{
IPUMS: data can be accessed at \url{https://www.ipums.org/}
}
\usage{
sanfrancisco_sample
}
\description{
A pre-loaded dataset for the population of San Francisco, CA
}
\keyword{datasets}
