% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_people.R
\name{create_dots}
\alias{create_dots}
\title{Generate dataframe with locations of dots representing people on map}
\usage{
create_dots(data, num_people = 100)
}
\arguments{
\item{data}{A dataframe containing shapefiles to plot census tract boundaries for each year, output of get_data_wide function}

\item{num_people}{number of individuals that each dot represents}
}
\description{
Generate dataframe with locations of dots representing people on map
}
