% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{get_data_long}
\alias{get_data_long}
\alias{get_data_wide}
\title{Generate data frame with demographic information for 1950-2020 for specified region}
\usage{
get_data_long(path, state, county)

get_data_wide(path, state, county)
}
\arguments{
\item{path}{character string speciying the path where data is located}

\item{state}{character string specifying the name of the desired state}

\item{county}{character string specifying the name of the desired county}
}
\description{
Generate data frame with demographic information for 1950-2020 for specified region

Generate data frame used to map demographics
}
